# src/main/java/com/probestack/forgestudio/design/api/CartsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.Cart;
import com.probestack.forgestudio.design.model.CartListResponse;
import com.probestack.forgestudio.design.model.CreateCartRequest;
import com.probestack.forgestudio.design.model.ErrorResponse;
import com.probestack.forgestudio.design.model.UpdateCartRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-13T10:33:33.240785079Z[GMT]")@Validated
@Tag(name = "Carts", description = "Operations on shopping carts")
public interface CartsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /carts : Create a new cart
     * Creates an empty cart for a customer or guest session. A customer may hold only one ACTIVE cart at a time.
     *
     * @param createCartRequest  (required)
     * @return Cart created successfully (status code 201)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or An active cart already exists for this customer (status code 409)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "createCart",
        summary = "Create a new cart",
        description = "Creates an empty cart for a customer or guest session. A customer may hold only one ACTIVE cart at a time.",
        tags = { "Carts" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Cart created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Cart.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "An active cart already exists for this customer", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/carts",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Cart> createCart(
        @Parameter(name = "CreateCartRequest", description = "", required = true) @Valid @RequestBody CreateCartRequest createCartRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"couponCodes\" : [ \"SUMMER10\" ], \"discountTotal\" : 9.0, \"grandTotal\" : 80.98, \"subtotal\" : 89.98, \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"c1a2r3t4\", \"sessionId\" : \"sess-9f8e7d\", \"state\" : \"ACTIVE\", \"items\" : [ { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" }, { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" } ], \"expiresAt\" : \"2026-08-19T10:00:00Z\", \"updatedAt\" : \"2026-08-12T10:00:00Z\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * DELETE /carts/{cartId} : Empty a cart
     * Removes all line items and releases any soft stock reservations. The cart itself is retained.
     *
     * @param cartId Unique cart identifier (UUID) (required)
     * @return Cart emptied successfully (no content returned) (status code 204)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "emptyCart",
        summary = "Empty a cart",
        description = "Removes all line items and releases any soft stock reservations. The cart itself is retained.",
        tags = { "Carts" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Cart emptied successfully (no content returned)"),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/carts/{cartId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> emptyCart(
        @Parameter(name = "cartId", description = "Unique cart identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("cartId") String cartId
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }


    /**
     * GET /carts/{cartId} : Get a cart by ID
     * Retrieves the cart with its line items, applied coupons, and calculated totals.
     *
     * @param cartId Unique cart identifier (UUID) (required)
     * @return Cart found (status code 200)
     *         or Cart not found (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "getCartById",
        summary = "Get a cart by ID",
        description = "Retrieves the cart with its line items, applied coupons, and calculated totals.",
        tags = { "Carts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Cart found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Cart.class))
            }),
            @ApiResponse(responseCode = "404", description = "Cart not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/carts/{cartId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Cart> getCartById(
        @Parameter(name = "cartId", description = "Unique cart identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("cartId") String cartId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"couponCodes\" : [ \"SUMMER10\" ], \"discountTotal\" : 9.0, \"grandTotal\" : 80.98, \"subtotal\" : 89.98, \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"c1a2r3t4\", \"sessionId\" : \"sess-9f8e7d\", \"state\" : \"ACTIVE\", \"items\" : [ { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" }, { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" } ], \"expiresAt\" : \"2026-08-19T10:00:00Z\", \"updatedAt\" : \"2026-08-12T10:00:00Z\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /carts : List all carts
     * Returns a paginated list of carts, optionally filtered by customer or abandonment state.
     *
     * @param customerId Filter by the owning customer (optional)
     * @param state Filter by cart lifecycle state (optional)
     * @param page Page number (starts at 1) (optional, default to 1)
     * @param size Number of items per page (optional, default to 20)
     * @return Successful response (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "listCarts",
        summary = "List all carts",
        description = "Returns a paginated list of carts, optionally filtered by customer or abandonment state.",
        tags = { "Carts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartListResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/carts",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CartListResponse> listCarts(
        @Parameter(name = "customerId", description = "Filter by the owning customer", in = ParameterIn.QUERY) @Valid @RequestParam(value = "customerId", required = false) String customerId,
        @Parameter(name = "state", description = "Filter by cart lifecycle state", in = ParameterIn.QUERY) @Valid @RequestParam(value = "state", required = false) String state,
        @Min(1) @Parameter(name = "page", description = "Page number (starts at 1)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "1") Integer page,
        @Min(1) @Max(100) @Parameter(name = "size", description = "Number of items per page", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"size\" : 20, \"totalPages\" : 26, \"page\" : 1, \"content\" : [ { \"couponCodes\" : [ \"SUMMER10\" ], \"discountTotal\" : 9.0, \"grandTotal\" : 80.98, \"subtotal\" : 89.98, \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"c1a2r3t4\", \"sessionId\" : \"sess-9f8e7d\", \"state\" : \"ACTIVE\", \"items\" : [ { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" }, { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" } ], \"expiresAt\" : \"2026-08-19T10:00:00Z\", \"updatedAt\" : \"2026-08-12T10:00:00Z\" }, { \"couponCodes\" : [ \"SUMMER10\" ], \"discountTotal\" : 9.0, \"grandTotal\" : 80.98, \"subtotal\" : 89.98, \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"c1a2r3t4\", \"sessionId\" : \"sess-9f8e7d\", \"state\" : \"ACTIVE\", \"items\" : [ { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" }, { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" } ], \"expiresAt\" : \"2026-08-19T10:00:00Z\", \"updatedAt\" : \"2026-08-12T10:00:00Z\" } ], \"totalElements\" : 512 }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * PUT /carts/{cartId} : Replace cart contents
     * Replaces the full set of line items and coupons. The &#x60;customerId&#x60; cannot be changed once set.
     *
     * @param cartId Unique cart identifier (UUID) (required)
     * @param updateCartRequest  (required)
     * @return Cart updated successfully (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "updateCart",
        summary = "Replace cart contents",
        description = "Replaces the full set of line items and coupons. The `customerId` cannot be changed once set.",
        tags = { "Carts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Cart updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Cart.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/carts/{cartId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Cart> updateCart(
        @Parameter(name = "cartId", description = "Unique cart identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("cartId") String cartId,
        @Parameter(name = "UpdateCartRequest", description = "", required = true) @Valid @RequestBody UpdateCartRequest updateCartRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"couponCodes\" : [ \"SUMMER10\" ], \"discountTotal\" : 9.0, \"grandTotal\" : 80.98, \"subtotal\" : 89.98, \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"c1a2r3t4\", \"sessionId\" : \"sess-9f8e7d\", \"state\" : \"ACTIVE\", \"items\" : [ { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" }, { \"unitPrice\" : 44.99, \"quantity\" : 2, \"sku\" : \"SKU-12345\", \"productName\" : \"Wireless Mouse\" } ], \"expiresAt\" : \"2026-08-19T10:00:00Z\", \"updatedAt\" : \"2026-08-12T10:00:00Z\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}